Spawning Actors

+++ sample_project/Source/sample_project/Stream Layer/DeadReckoning.h
@@ -20,7 +20,7 @@ public:
 	static TArray<double> DeadReckoningPoint(double speed, double timespan, TArray<double> currentPoint, double headingDegrees);
 	static TArray<double> MoveByDistanceAndHeading(TArray<double> currentPoint, double distanceMeters, double headingDegrees);
 	static double InitialBearingTo(TArray<double> orgPoint, TArray<double> dstPoint);
-	static double Wrap360(double degrees);
+	static double Wrap360(int degrees);
 	static TArray<double> MoveTowards(TArray<double> targetLocation, TArray<double> currentLocation, double maxDistanceDelta);
 	
 	inline static double earthRadiusMeters = 6356752.3142;
